pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unravel-agent"
version = "0.1.0"
description = "A DSPy-powered agent that solves the Unravel.tech hiring riddle"
requires-python = ">=3.11"
dependencies = [
    "dspy>=2.6.0",
    "playwright>=1.44.0",
    "pdfplumber>=0.11.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
